Refactor keyboard input
Make function for low-level keypress reading, and another function for mapping keypresses to editor operations.

Stop printing out keypresses to the screen.

editorProcessKeypress() waits for a keypress, then handles it.

editorReadKey() waits for one keypress and returns it. It belongs in the /*** terminal ***/ section of the file since it deals with low-level terminal input, while editorProcessKeypress() is a higher-level function that maps keypresses to editor operations. It belongs in the /*** input ***/ section of the file.

This creates a vastly simplified main() function.