Clear the screen
Render editor's user interface to the screen after each keypress.

write() and STDOUT_FILENO come from <unistd.h>.

The 4 in the write() call is the number of bytes we want to write to the terminal. The first byte is the escape character \x1b, which is the escape character, or 27 in decimal. The other three bytes are [2J. The [ is an escape sequence introducer, and the J is a command letter that tells the terminal to clear the screen. The 2 is an argument for the J command that tells the terminal to clear the entire screen.