Add editorDrawRows function
This function will draw the rows of the buffer of text being edited. At this point it draws a tilde in each row, which means that row is not part of the file and can't contain text.

After done drawing, another <esc>[H escape sequence is sent to the terminal to move the cursor back to the top left corner of the screen.